delivery packaging delete fix.

+++ aon-occam/src/main/java/com/esferalia/aon/occam/impl/jooq/dao/PackagingDAO.java
@@ -514,43 +514,45 @@ public class PackagingDAO {
 		
 		Item containerItem = ItemDAO.get(ctx, packaging.getContainer().getItem(), new Options().setFull(true));
 		containerItem.getItemComposition().stream().forEach(itemComposition -> {
-			Integer productId = itemComposition.getComposition().getProduct().getId();
-			DeliveryDetail dd = DeliveryDetailDAO.get(ctx, f -> f.getDelivery().eq(deliveryId)
-				.and(f.getItem().eq(itemComposition.getCompositionItemId())));
-			double q = dd.getQuantity() - itemComposition.getQuantity();
-			if(q == 0) {
-				DeliveryDetailDAO.delete(ctx, f -> f.getIdProperty().eq(dd.getId()));
-			} else {
-				dd.setQuantity(q);
-				DeliveryDetailDAO.save(ctx, dd);
+			if(itemComposition.getId() != null) {
+				Integer productId = itemComposition.getComposition().getProduct().getId();
+				DeliveryDetail dd = DeliveryDetailDAO.get(ctx, f -> f.getDelivery().eq(deliveryId)
+					.and(f.getItem().eq(itemComposition.getCompositionItemId())));
+				double q = dd.getQuantity() - itemComposition.getQuantity();
+				if(q == 0) {
+					DeliveryDetailDAO.delete(ctx, f -> f.getIdProperty().eq(dd.getId()));
+				} else {
+					dd.setQuantity(q);
+					DeliveryDetailDAO.save(ctx, dd);
+				}
+				SalesDetail sd = SalesDetailDAO.get(ctx, f -> f.getIdProperty().eq(dd.getSalesDetail()));
+				sd.setDelivered(sd.getDelivered() - itemComposition.getQuantity());
+				if(bool && sd.getDelivery() != null && sd.getDelivered() == 0.0 && sd.getDelivery().equals(deliveryId))
+					sd.setDelivery(null);
+				sd.setStatus(sd.getDelivered() > 0 ? SalesDetailStatus.PARTIAL_SETTLED : SalesDetailStatus.PENDING);
+				SalesDetailDAO.save(ctx, sd);
+				
+				// TODO FALTA REVISAR PEDIDO!!!
+				List<SalesDetail> list = SalesDetailDAO.getStream(ctx, f -> 
+					f.getSalesProperty().eq(sd.getSales().getId())
+					.and(f.getDomainProperty().eq(sd.getDomain()))
+					.and(f.getDeliveryProperty().isNotNull()))
+					.toList();
+				if(list.isEmpty()) {
+					Sales sales = SalesDAO.get(ctx, sd.getSales().getId());
+					sales.setStatus(SalesStatus.PENDING);
+					SalesDAO.save(ctx, sales);
+				}
+				
+				Stock stock = WarehouseDAO.getStock(ctx, f -> f.getDomainProperty().eq(ctx.getDomainId())
+					.and(f.getItemProperty().eq(itemComposition.getCompositionItemId())));
+				if(stock.getId() != null) {
+					stock.setQuantity(stock.getQuantity() + itemComposition.getQuantity());
+					WarehouseDAO.saveStock(ctx, stock);
+				}
+				
+				deleteItemBox(ctx, deliveryId, productId, itemComposition.getQuantity());	
 			}
-			SalesDetail sd = SalesDetailDAO.get(ctx, f -> f.getIdProperty().eq(dd.getSalesDetail()));
-			sd.setDelivered(sd.getDelivered() - itemComposition.getQuantity());
-			if(bool && sd.getDelivery() != null && sd.getDelivered() == 0.0 && sd.getDelivery().equals(deliveryId))
-				sd.setDelivery(null);
-			sd.setStatus(sd.getDelivered() > 0 ? SalesDetailStatus.PARTIAL_SETTLED : SalesDetailStatus.PENDING);
-			SalesDetailDAO.save(ctx, sd);
-			
-			// TODO FALTA REVISAR PEDIDO!!!
-			List<SalesDetail> list = SalesDetailDAO.getStream(ctx, f -> 
-				f.getSalesProperty().eq(sd.getSales().getId())
-				.and(f.getDomainProperty().eq(sd.getDomain()))
-				.and(f.getDeliveryProperty().isNotNull()))
-				.toList();
-			if(list.isEmpty()) {
-				Sales sales = SalesDAO.get(ctx, sd.getSales().getId());
-				sales.setStatus(SalesStatus.PENDING);
-				SalesDAO.save(ctx, sales);
-			}
-			
-			Stock stock = WarehouseDAO.getStock(ctx, f -> f.getDomainProperty().eq(ctx.getDomainId())
-				.and(f.getItemProperty().eq(itemComposition.getCompositionItemId())));
-			if(stock.getId() != null) {
-				stock.setQuantity(stock.getQuantity() + itemComposition.getQuantity());
-				WarehouseDAO.saveStock(ctx, stock);
-			}
-			
-			deleteItemBox(ctx, deliveryId, productId, itemComposition.getQuantity());
 		});
 		
 		// QUITAR envase en delivery detail.
